+++ wagtailsharing/views.py
@@ -95,7 +95,10 @@ class ServeView(View):
                 return result
 
         # Get the latest revision for the requested page.
-        page = page.get_latest_revision_as_page()
+        if WAGTAIL_VERSION >= (4, 0):
+            page = page.get_latest_revision_as_object()
+        else:
+            page = page.get_latest_revision_as_page()
 
         # Call the before_serve_shared_page hook.
         for fn in hooks.get_hooks("before_serve_shared_page"):